Tests/LibWeb/Text/expected/WebAudio/OfflineAudioContext-startRendering.txt

after-start:running
statechange:running
after-await:closed:1
InvalidStateError: Rendering is already started
complete:closed
